+++ idea/testData/refactoring/rename/renameClassTypeParameterRedeclarationWithOuterScope/before/test.kt
@@ -1,0 +1,3 @@
+class A<T> {
+    inner class B</*rename*/U, V>
+}